Adjusted default value at SshVersion config element

+++ FeatureBranches/Putty/Source/Terminals.Common/Configuration/FavoriteConfigurationElement.cs
@@ -652,7 +652,7 @@ namespace Terminals
             }
         }
 
-        [ConfigurationProperty("sshVersion", IsRequired = true, DefaultValue = 0)]
+        [ConfigurationProperty("sshVersion", IsRequired = true, DefaultValue = (byte)0)]
         public byte SshVersion {
             get {
                 return (byte)this["sshVersion"];
